requirements.txt
pytest
discord.py
python-dotenv
quart
Pillow